# swagger/v2/swagger.yaml
swagger: '2.0'
info:
  title: METAX API
  version: v2
host: __METAX_ENV_DOMAIN__
schemes:
  - https
basePath: ''
produces:
  - application/json
paths:
...
  /rest/v2/files:
    get:
      summary: Get list of files
      parameters:
        - name: project_identifier
          in: query
          description: Filter files by project
          required: false
          type: string
        - name: file_path
          in: query
          description: Filters files by path. Requires query parameter project_identifier.
          required: false
          type: string
        - name: offset
          in: query
          description: Offset for paging
          required: false
          type: integer
        - name: limit
          in: query
          description: Limit for paging
          required: false
          type: integer
          default: 10
        - name: ordering
          in: query
          description: Specify ordering of results by fields. Accepts a list of field names separated by a comma. Ordering can be reversed by prefixing field name with a '-' char.
          required: false
          type: string
        - name: fields
          in: query
          description: |
            Comma separated list of fields that is returned. Checksum -field can be further selected with separating the inner key
            with colon. Example: ?fields=checksum:value.
          required: false
          type: string
      responses:
        "200":
          description: Successful operation, return list of files
          schema:
            $ref: '#/definitions/FileList'
        '404':
          description: Not found. Also when not authenticated request
      tags:
        - File API
    post:
      summary: Create new file metadata
      consumes:
        - application/json
      parameters:
      - in: "body"
        name: "body"
        description: "Object or a list of objects to create."
        required: true
        schema:
          $ref: '#/definitions/File'
      - $ref: "#/parameters/dryrun"
      responses:
        '201':
          description: Returns the created object, or if a list was given, a list of objects and errors.
        '400':
          description: Parameters contained errors, response includes details.
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - File API
    put:
      summary: Bulk update
      consumes:
        - application/json
      parameters:
      - in: "body"
        name: "body"
        description: A list of objects to update.
        required: true
        schema:
          $ref: '#/definitions/FileList'
      - $ref: "#/parameters/allowed_projects"
      - $ref: "#/parameters/dryrun"
      responses:
        '200':
          description: Successful operation. Return values include a list of errors, if any.
        '400':
          description: All updates failed. A list of errors is returned.
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - File API
    patch:
      summary: Bulk update partial
      description: |
        The payload must include a field that can be used to identify the resource being updated. Acceptable identifier fields are: <b>id, identifier</b>
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: A list of (partial) objects to update.
          required: true
          schema:
            $ref: '#/definitions/FileList'
        - $ref: "#/parameters/allowed_projects"
        - $ref: "#/parameters/dryrun"
      responses:
        '200':
          description: Some or all objects were updated. Return values contain list of full updated objects, and may include a list of errors.
        '400':
          description: All updates failed. A list of errors is returned.
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - File API
    delete:
      summary: Bulk delete
      description: Mark files as deleted en masse.
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: A list of ids, or a list of identifiers.
          required: true
          schema:
            $ref: '#/definitions/StringList'
        - $ref: "#/parameters/dryrun"
      responses:
        '200':
          description: |
            Successful operation. At least some requested files were marked deleted. If some of the provided
            identifiers were not found, those are ignored. Returns count of deleted files in json body.
        '400':
          description: All updates failed. A list of errors is returned.
        '403':
          description: Forbidden. Must have permission for resource
        '404':
          description: None of the provided identifiers were found.
      tags:
        - File API
...
  /rest/v2/datasets/identifiers:
    get:
      summary: List all dataset identifiers
      parameters:
        - name: latest
          in: query
          description: Only return latest versions
          required: false
          type: boolean
        - name: owner_id
          in: query
          description: Id of the person who owns the record in metax
          required: false
          type: string
        - name: user_created
          in: query
          description: Id of the person who created the record in metax
          required: false
          type: string
        - name: curator
          in: query
          description: Curator identifier (field research_dataset-> curator-> identifier)
          required: false
          type: string
        - name: state
          in: query
          description: TPAS state (field preservation_state). Multiple states using OR-logic are queriable in the same request, e.g. state=5,6. See valid values from http://iow.csc.fi/model/mrd/CatalogRecord/ field preservation_state
          required: false
          type: string
      responses:
        "200":
          description: Successful operation, returns a list of all dataset metadata version identifiers.
      tags:
        - Dataset API
  /rest/v2/datasets/unique_preferred_identifiers:
    get:
      summary: List all unique dataset preferred identifiers
      parameters:
        - name: latest
          in: query
          description: Only return latest dataset versions
          required: false
          type: boolean
        - name: owner_id
          in: query
          description: Id of the person who owns the record in metax
          required: false
          type: string
        - name: user_created
          in: query
          description: Id of the person who created the record in metax
          required: false
          type: string
        - name: curator
          in: query
          description: Curator identifier (field research_dataset-> curator-> identifier)
          required: false
          type: string
        - name: state
          in: query
          description: TPAS state (field preservation_state). Multiple states using OR-logic are queriable in the same request, e.g. state=5,6. see valid values from http://iow.csc.fi/model/mrd/CatalogRecord/ field preservation_state
          required: false
          type: string
      responses:
        "200":
          description: Successful operation, returns a list of all unique dataset preferred identifiers.
      tags:
        - Dataset API
  /rest/v2/datasets/list:
    post:
      summary: List datasets corresponding IDs or identifiers given in body
      parameters:
        - name: body
          in: body
          description: A list of dataset IDs (integers) OR a list of datasets identifiers (strings). Returns datasets corresponding queried IDs.
          required: true
          schema:
            $ref: '#/definitions/StringList'
        - $ref: "#/parameters/fields"
      responses:
        "200":
          description: Successful operation, returns a list of queried datasets.
        "400":
          description: Received empty list of identifiers
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - Dataset API
...
  /rest/v2/datasets/{PID}/files:
    get:
      summary: Get list of all files in a dataset and their technical metadata.
      description: Get list of all files in a dataset and their technical metadata. If dataset is under embargo, or otherwise dataset access_type is not fully open, it is possible that the file list can not be retrieved.
      parameters:
        - name: PID
          in: path
          description: Persistent ID of dataset OR the internal pk
          required: true
          type: string
        - name: file_fields
          in: query
          description: Comma-separated list of field names to retrieve for files
          required: false
          type: string
        - name: removed_files
          in: query
          description: Return only deleted files
          required: false
          type: boolean
      responses:
        '200':
          description: Successful operation, return list of full files metadata.
          schema:
            $ref: '#/definitions/FileList'
      tags:
        - Dataset API
    post:
      summary: Add files to a dataset or exclude files from a dataset.
      description: |
        Add files to a dataset or exclude files from an existing dataset. This API is the sole means of adding or excluding files once a dataset has been created.
        - To add files, only the field "identifier" is required. To exclude fields, only the fields "identifier" and "exclude" are required (i.e. "exclude=true").
        - In general, this API can be used only for datasets that are in draft state, where files can be freely added or excluded.
        - When a dataset has been published, this API can be used in two situations only: 1) Dataset has never had any files added to it. Files can be added to the dataset once, but after that, adding or excluding files will be prevented. 2) Dataset is cumulative. Files can be freely added, but not excluded.
        - The request body can include also user metadata for files and directories that are added.
        - The user must be a member of the project where files are added from.
        - API only usable by the owner of the dataset.
      parameters:
        - name: PID
          in: path
          description: Persistent ID of dataset OR the internal pk
          required: true
          type: string
        - in: body
          name: body
          description: List of files or entire directories to include or exclude.
          required: true
          schema:
            $ref: '#/definitions/AddOrExcludeFiles'
      responses:
        '200':
          description: Successful operation, return number of files added and removed.
          schema:
            type: object
            properties:
              files_added:
                type: integer
                example: 3
              files_removed:
                type: integer
                example: 2
        '400':
          description: Operation resulted in an error, response includes details.
        '403':
          description: Forbidden. Must have permission for resource.
      tags:
        - Dataset API
...
  /rest/v2/datasets/{PID}/metadata_versions:
    get:
      summary: List old research_dataset entries of a record
      parameters:
        - name: PID
          in: path
          description: Persistent ID of dataset OR the internal pk
          required: true
          type: string
      responses:
        "200":
          description: Successful operation, return a list of entries. May return an empty list.
          schema:
            $ref: '#/definitions/StringList'
      tags:
        - Dataset API
...
parameters:
  dryrun:
    name: dryrun
    in: query
    description: Execute the operation normally, returning the same response as normally, but do not save anything to db, or publish anything to remote services.
    required: false
    type: boolean
  allowed_projects:
    name: allowed_projects
    in: query
    description: Ensures that content is part of intended projects.
    required: false
    type: string
...
  fields:
    name: fields
    in: query
    description: Comma separated list of fields that is returned. Note that nested fields are not supported.
    required: false
    type: string
definitions:
...
  AddOrExcludeFiles:
    type: object
    additionalProperties: false
    minProperties: 1
    properties:
      files:
        type: array
        items:
          type: object
          minProperties: 1
          anyOf:
            - $ref: 'https://__METAX_ENV_DOMAIN__/apischemas/v2/dataset_files_schema.json#/definitions/FileAdditionEvent'
            - $ref: 'https://__METAX_ENV_DOMAIN__/apischemas/v2/dataset_files_schema.json#/definitions/FileMetadata'
      directories:
        type: array
        items:
          type: object
          minProperties: 1
          anyOf:
            - $ref: 'https://__METAX_ENV_DOMAIN__/apischemas/v2/dataset_files_schema.json#/definitions/DirectoryAdditionEvent'
            - $ref: 'https://__METAX_ENV_DOMAIN__/apischemas/v2/dataset_files_schema.json#/definitions/DirectoryMetadata'
    type: object
    properties:
      files:
        type: array
        items:
          $ref: '#/definitions/ResearchDatasetFile'
      directories:
        type: array
        items:
          $ref: '#/definitions/ResearchDatasetDirectory'
...
  File:
    $ref: 'https://__METAX_ENV_DOMAIN__/apischemas/v2/file.json#/definitions/File'
...
  StringList:
    type: array
    items:
      type: string
      example: "value"
...
  FileList:
    type: array
    items:
      $ref: '#/definitions/File'
...